Fix grub-install bug
`grub-install` at the end of phase4/grub.sh was erroring out
with complaints about the loop device and suggesting to check
the device.map file.

LFS has the reader run `grub-install` at the END of the build,
but because that command was included in the phase4/grub.sh script,
this meant it was executed while there were still more packages
to build. Moving the phase4/grub.sh build to the end fixed this issue.